pyproject.toml
[project]
name = "robot-arm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "faster-whisper>=1.2.0",
    "imutils>=0.5.4",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "onnxruntime>=1.22.1",
    "openai-whisper>=20250625",
    "opencv-python-headless>=4.11.0.86",
    "paddleocr>=3.2.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-levenshtein>=0.27.1",
    "tensorflow>=2.20.0",
]
